step-wise force data finished

+++ forceAnalysis/operations/gopro_audio_force_matching.py
@@ -12,12 +12,12 @@ from scipy import signal, interpolate
 import matplotlib.pyplot as plt
 
 
-def extract_force_details(vline_list, df_step4, df_forces, date, i):
+def extract_force_details(vline_list, df_step4, df_forces, date, i, gait):
     """
     This function is called from within the for loop iterating through the individual audio files in match_audio_and_force()
     to extract max, min, and mean from the individual step intervals.
     This data is stored as a new data sheet (csv saved from df_force_detail) containing:
-    "audiofile", "forcefile", "run", "foot_on_fp", "foot", "step", "step_forceframes_start",
+    "audiofile", "forcefile", "gait", "run", "foot_on_fp", "foot", "step", "step_forceframes_start",
     "step_forceframes_end", "Fx_max", "Fy_max", "Fz_max", "Fx_min", "Fy_min", "Fz_min",
     "Fx_mean", "Fy_mean", "Fz_mean"
 
@@ -29,10 +29,10 @@ def extract_force_details(vline_list, df_step4, df_forces, date, i):
     :return: df_rows_append: df containing the rows of step intervals for current run i to append
     """
 
-    print(f"\n EXTRACTING STEP WISE FORCE DATA for step4 row {i}...\n")
+    print(f"\n EXTRACTING STEP WISE FORCE DATA for step4.csv row {i}...\n")
 
     # creates empty DataFrame with the columns to be filled in with step-wise info. Data for each step will be added to this.
-    df_rows_append = pd.DataFrame(columns=["audiofile", "forcefile", "run", "foot_on_fp", "foot", "step", "step_forceframes_start",
+    df_rows_append = pd.DataFrame(columns=["audiofile", "forcefile", "gait", "run", "foot_on_fp", "foot", "step", "step_forceframes_start",
     "step_forceframes_end", "Fx_max", "Fy_max", "Fz_max", "Fx_min", "Fy_min", "Fz_min",
     "Fx_mean", "Fy_mean", "Fz_mean"])
 
@@ -56,39 +56,41 @@ def extract_force_details(vline_list, df_step4, df_forces, date, i):
             Fy_mean = np.nanmean(df_forces["Fy"][vline_list[j-1]:vline_list[j]])
             Fz_mean = np.nanmean(df_forces["Fz"][vline_list[j-1]:vline_list[j]])
 
+            print(f"Fx_mean: {Fx_mean}, Fy_mean: {Fy_mean}, Fz_mean: {Fz_mean}")
+
             if j == 1:
-                foot_on_fp = df_step4["foot_on_fp"][i]
+                foot_on_fp = int(df_step4["foot_on_fp"][i])
 
             step = foot_on_fp + (j-1)
 
             if foot_on_fp == 4:
-                foot = "FR"
-            elif foot_on_fp == 5:
                 foot = "HR"
+            elif foot_on_fp == 5:
+                foot = "FR"
             else:
                 foot = np.nan
 
             #create single DataFrame column to return and append & assign values to all other columns for current step interval (j-1):j
             run = df_step4["run"][i]
             df_row_append = pd.DataFrame({
-                "audiofile": df_step4["audiofile"][i],
-                "forcefile": f"{date}_run{run}.txt",
-                "run": run,
-                "foot_on_fp": foot_on_fp,
-                "foot": foot,
-                "step": step,
-                "step_forceframes_start": vline_list[j-1],
-                "step_forceframes_end": vline_list[j],
-                "Fx_max": Fx_max,
-                "Fy_max": Fy_max,
-                "Fz_max": Fz_max,
-                "Fx_min": Fx_min,
-                "Fy_min": Fy_min,
-                "Fz_min": Fz_min,
-                "Fx_mean": Fx_mean,
-                "Fy_mean": Fy_mean,
-                "Fz_mean": Fz_mean
-            })
+                "audiofile": [df_step4["audiofile"][i]],
+                "forcefile": [f"{date}_run{run}.txt"],
+                "run": [run],
+                "gait": [gait],
+                "foot_on_fp": [foot_on_fp],
+                "foot": [foot],
+                "step": [step],
+                "step_forceframes_start": [vline_list[j-1]],
+                "step_forceframes_end": [vline_list[j]],
+                "Fx_max": [Fx_max],
+                "Fy_max": [Fy_max],
+                "Fz_max": [Fz_max],
+                "Fx_min": [Fx_min],
+                "Fy_min": [Fy_min],
+                "Fz_min": [Fz_min],
+                "Fx_mean": [Fx_mean],
+                "Fy_mean": [Fy_mean],
+                "Fz_mean": [Fz_mean]})
         df_rows_append = df_rows_append.append(df_row_append, ignore_index=True)
 
     print("DataFrame with step-wise data to append: \n", df_rows_append)
@@ -96,7 +98,7 @@ def extract_force_details(vline_list, df_step4, df_forces, date, i):
     return df_rows_append
 
 
-def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audio_files, date):
+def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audio_files, date, gait):
     """
     this function will use the {date}_gammaForces_step3.csv file and the dict containing the audio peaks from the
     gopro_audio_analysis module.
@@ -125,7 +127,7 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
     pretrigger_samples = 90000
 
     ### Create data frame for step-wise force extraction:
-    df_force_detail_colnames = ["audiofile", "forcefile", "run", "foot_on_fp", "foot", "step", "step_forceframes_start",
+    df_force_detail_colnames = ["audiofile", "forcefile", "gait", "run", "foot_on_fp", "foot", "step", "step_forceframes_start",
                                 "step_forceframes_end", "Fx_max", "Fy_max", "Fz_max", "Fx_min", "Fy_min", "Fz_min",
                                 "Fx_mean", "Fy_mean", "Fz_mean"]
     df_force_detail = pd.DataFrame(columns=df_force_detail_colnames)
@@ -144,17 +146,17 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
     # check if column "foot_on_fp" exists:
     if "foot_on_fp" in gammaForces_columns:
         # iterate through the runs from this date:
-        for i in range(len(df_gammaForces["audiofile"])):
-            audio_file_name = df_gammaForces.loc[i, "audiofile"]
+        for row in range(len(df_gammaForces["audiofile"])):
+            audio_file_name = df_gammaForces.loc[row, "audiofile"]
 
             # get the refined status for the run and only proceed if green (all peaks in audio detected)
-            refined_satus = df_gammaForces.loc[i, "status_refined"]
+            refined_satus = df_gammaForces.loc[row, "status_refined"]
             if refined_satus == "green":
                 print(f"\n --- matching {audio_file_name} to force data...")
 
                 # find the corresponding filename in list of current audio_file_name
                 audio_file_path = [i for i in l_gopro_audio_files if audio_file_name in i][0]
-                audio_framerate = df_gammaForces.loc[i, "audio_framerate"]
+                audio_framerate = df_gammaForces.loc[row, "audio_framerate"]
 
                 ### READ IN THE AUDIO FILE:
                 raw = wave.open(audio_file_path)
@@ -163,8 +165,8 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
                 signal = np.frombuffer(signal, dtype="int16")
                 # cut the audio signal to step peaks interval:
                 # get time start and end of step spikes interval
-                audio_interval_start_s = df_gammaForces.loc[i, "audio_wave_start_s"]
-                audio_interval_end_s = df_gammaForces.loc[i, "audio_wave_end_s"]
+                audio_interval_start_s = df_gammaForces.loc[row, "audio_wave_start_s"]
+                audio_interval_end_s = df_gammaForces.loc[row, "audio_wave_end_s"]
                 # convert to frames
                 f_rate = raw.getframerate()
                 f_bytes = raw.getsampwidth()
@@ -179,14 +181,14 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
                 print(f"length of audio data: {length_audio}")
 
                 ### READ IN RESPECTIVE FORCE FILE for current run i (tab delimited) and add column names
-                force_file = date + "_run" + str(df_gammaForces.loc[i, "run"]) + ".txt"
+                force_file = date + "_run" + str(df_gammaForces.loc[row, "run"]) + ".txt"
                 force_file_path = os.path.join(path_gammaForces_sheet, force_file)
                 forces_gamma_columnnames = ["Fx", "Fy", "Fz", "Tx", "Ty", "Tz"]
                 df_force_file = pd.read_csv(force_file_path, sep="\t", names=forces_gamma_columnnames)
                 length_force = df_force_file['Fx'].count()
                 print(f"length of force data: {length_force}")
 
-                foot_on_fp = int(df_gammaForces.loc[i, "foot_on_fp"])
+                foot_on_fp = int(df_gammaForces.loc[row, "foot_on_fp"])
                 print(f"foot on force plate: {foot_on_fp}")
                 if foot_on_fp == 4:
                     plot_title = "FR on Force Plate"
@@ -242,12 +244,12 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
                     # to each element in the list add the sum of all previous elements, otherwise all hlines will be at a
                     # similar spot as only the relative difference is plotted:
                     absolute_audio_peaks = []
-                    for i, item in enumerate(diff_relevant_audio_peak_frames):
-                        item_new = sum(diff_relevant_audio_peak_frames[:i+1])
+                    for k, item in enumerate(diff_relevant_audio_peak_frames):
+                        item_new = sum(diff_relevant_audio_peak_frames[:k+1])
                         # print(item, item_new)
                         absolute_audio_peaks.append(item_new)
                     print("absolute_audio_peaks: ", absolute_audio_peaks)
-                    audio_peaks_in_force_frames = [(i+ind_min_z_force) for i in absolute_audio_peaks]
+                    audio_peaks_in_force_frames = [(l+ind_min_z_force) for l in absolute_audio_peaks]
                     print("plot_vlines_at...: ", audio_peaks_in_force_frames)
 
                     # plot forces:
@@ -263,16 +265,16 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
                     plt.plot(xnew[limits[0]:limits[1]], ynew_force_z[limits[0]:limits[1]], color='red', alpha=0.5,
                              label="Fz")
                     plt.scatter(ind_min_z_force, min_z_force)
-                    for i, audio_p in enumerate(audio_peaks_in_force_frames):
+                    for u, audio_p in enumerate(audio_peaks_in_force_frames):
                         plt.vlines(audio_p, min(ynew_force_z), max(ynew_force_z))
-                        plt.text(audio_p + audio_p/100, max(ynew_force_z), f"step {foot_on_fp + i}")
+                        plt.text(audio_p + audio_p/100, max(ynew_force_z), f"step {foot_on_fp + u}")
                     plt.title(f"{audio_file_name} - {plot_title} - interp")
                     plt.savefig(os.path.join(path_gammaForces_sheet, "plots",
                                              f"{audio_file_name}_forces_and_audiopeaks.jpg"))  # save as jpg
                     plt.show(block=True)
 
                     ## call step-wise force extraction function here:
-                    df_rows_append = extract_force_details(audio_peaks_in_force_frames, df_gammaForces, df_force_file, date, i)
+                    df_rows_append = extract_force_details(audio_peaks_in_force_frames, df_gammaForces, df_force_file, date, row, gait)
                     df_force_detail = df_force_detail.append(df_rows_append, ignore_index=True)
 
                 else:
@@ -289,7 +291,7 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
                     ind_min_z_force = np.argmin(df_force_file['Fz'])
                     print(f"z_force minimum: {min_z_force}, index: {ind_min_z_force}")
 
-                    foot_on_fp = int(df_gammaForces.loc[i, "foot_on_fp"])
+                    foot_on_fp = int(df_gammaForces.loc[row, "foot_on_fp"])
                     print(f"foot on force plate: {foot_on_fp}")
 
                     # find the audio peak for the n-th foot, which is the one on fp:
@@ -304,24 +306,24 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
                     diff_relevant_audio_peak_frames = np.insert(diff_relevant_audio_peak_frames, 0, 0)
 
                     # convert frame difference into time interval:
-                    diff_relevant_audio_peak_times = [i/f_rate for i in diff_relevant_audio_peak_frames]
+                    diff_relevant_audio_peak_times = [m/f_rate for m in diff_relevant_audio_peak_frames]
                     print("diff_relevant_audio_peak_times: ", diff_relevant_audio_peak_times)
 
                     ## now convert this to force frame differences:
                     force_sample_rate = 2500
-                    diff_audio_in_force_frames = [i*force_sample_rate for i in diff_relevant_audio_peak_times]
+                    diff_audio_in_force_frames = [b*force_sample_rate for b in diff_relevant_audio_peak_times]
                     print("diff_audio_in_force_frames: ", diff_audio_in_force_frames)
 
                     # to each element in the list add the sum of all previous elements, otherwise all hlines will be at a
                     # similar spot as only the relative difference is plotted:
                     absolute_force_audio_peaks = []
-                    for i, item in enumerate(diff_audio_in_force_frames):
-                        item_new = sum(diff_audio_in_force_frames[:i + 1])
+                    for it, item in enumerate(diff_audio_in_force_frames):
+                        item_new = sum(diff_audio_in_force_frames[:it + 1])
                         # print(item, item_new)
                         absolute_force_audio_peaks.append(item_new)
                     print("absolute_force_audio_peaks: ", absolute_force_audio_peaks)
 
-                    absolute_force_audio_peaks_plot = [(i + ind_min_z_force) for i in absolute_force_audio_peaks]
+                    absolute_force_audio_peaks_plot = [(d + ind_min_z_force) for d in absolute_force_audio_peaks]
                     print("plot_vlines_at...: ", absolute_force_audio_peaks_plot)
 
                     min_lim = 0
@@ -334,21 +336,24 @@ def match_audio_and_force(dict_audio_peaks, path_gammaForces_sheet, l_gopro_audi
                     plt.plot(x[limits[0]:limits[1]], df_force_file['Fz'][limits[0]:limits[1]], color='red', alpha=0.5,
                              label="Fz")
                     plt.scatter(ind_min_z_force, min_z_force)
-                    for i, audio_p in enumerate(absolute_force_audio_peaks_plot):
+                    for a, audio_p in enumerate(absolute_force_audio_peaks_plot):
                         plt.vlines(audio_p, min(list(df_force_file['Fx'])), max(list(df_force_file['Fx'])))
-                        plt.text(audio_p + audio_p/100, max(list(df_force_file['Fx'])), f"step {foot_on_fp + i}")
+                        plt.text(audio_p + audio_p/100, max(list(df_force_file['Fx'])), f"step {foot_on_fp + a}")
                     plt.title(f"{audio_file_name} - {plot_title}")
                     plt.savefig(os.path.join(path_gammaForces_sheet, "plots",
                                              f"{audio_file_name}_forces_and_audiopeaks_noInterp.jpg"))  # save as jpg
                     plt.show(block=True)
 
                     ## call step-wise force extraction function here:
-                    df_rows_append = extract_force_details(absolute_force_audio_peaks_plot, df_gammaForces, df_force_file, date, i)
+                    df_rows_append = extract_force_details(absolute_force_audio_peaks_plot, df_gammaForces, df_force_file, date, row, gait)
                     df_force_detail = df_force_detail.append(df_rows_append, ignore_index=True)
 
         #### STORE STEP-WISE FORCE DATA AS CSV FILE:
         force_detail_path = os.path.join(path_gammaForces_sheet, f"{date}_forceDetailsSteps.csv")
-        df_force_detail.to_csv(force_detail_path, index=False)
+        if os.path.exists(force_detail_path):
+            print(f"{date}_forceDetailsSteps.csv  already exists. To overwrite delete the current file in folder:\n{force_detail_path}")
+        else:
+            df_force_detail.to_csv(force_detail_path, index=False)
 
     else:
         print(error_message)
